+++ src/libnative/io/timer_timerfd.rs
@@ -28,8 +28,6 @@
 //!
 //! As with timer_other, all units in this file are in units of millseconds.
 
-#[allow(non_camel_case_types)];
-
 use std::comm::Data;
 use std::libc;
 use std::ptr;
